Split main_window to multiple model classes.
